# carepal-backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: CarePal HR Admin API
  version: 0.1.0
  description: |
    Internal API for the CarePal HR Admin tool. Manages hiring requisitions,
    candidates, interviews, headcount, and documents for CarePal Money's
    talent acquisition team.

    **Authentication:** the backend boots in one of two modes (controlled by
    the `AUTH_MODE` env var):

    - `mock` (dev/CI default): send `x-user-email` header with every
      request except `/health`. Missing or unknown emails return 401.
    - `google` (production default): send `Authorization: Bearer <id_token>`
      with a Google ID token. The backend verifies the token, enforces an
      email allowlist (carepalmoney.com, impactguru.com, plus one personal
      gmail), and auto-creates first-time users with role `ta`.

    **Authorization:** role-based. `admin` bypasses all role checks.
    Only `approver` (or `admin`) can create or approve requisitions.
servers:
  - url: http://localhost:4000
    description: Local dev (direct)
  - url: http://localhost:5173
    description: Local dev (via Vite proxy)
...
components:
  securitySchemes:
    mockAuth:
      type: apiKey
      in: header
      name: x-user-email
      description: Dev/CI mock auth — email of an existing user row
    googleIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Production auth — Google ID token (JWT) verified server-side
...
  schemas:
...
    Interview:
      type: object
      properties:
        id: { type: integer }
        candidateId: { type: string }
        round: { type: integer, enum: [1, 2] }
        interviewerName: { type: string }
        scheduledDate: { type: string, format: date }
        scheduledTime: { type: string, nullable: true }
        mode: { type: string, enum: [Virtual, In-Person] }
        locationOrLink: { type: string, nullable: true }
        result: { type: string, enum: [Select, Reject, No-show], nullable: true }
        cancelledAt: { type: string, format: date-time, nullable: true }
        cancelledReason: { type: string, nullable: true }
...
security:
  - mockAuth: []
  - googleIdToken: []
paths:
...
  /api/interviews:
    get:
      tags: [interviews]
      summary: List interviews (with filters)
      parameters:
        - { name: candidateId, in: query, schema: { type: string } }
        - { name: dateFrom, in: query, schema: { type: string, format: date }, description: Inclusive lower bound on scheduled_date }
        - { name: dateTo, in: query, schema: { type: string, format: date }, description: Inclusive upper bound on scheduled_date }
        - { name: round, in: query, schema: { type: integer, enum: [1, 2] } }
        - { name: result, in: query, schema: { type: string, enum: [Select, Reject, No-show, Scheduled] }, description: "'Scheduled' is a pseudo-value matching result IS NULL" }
        - { name: interviewerName, in: query, schema: { type: string } }
        - { name: includeCancelled, in: query, schema: { type: boolean }, description: When true, include soft-cancelled rows (default false) }
      responses:
        '200':
          description: Array of Interview
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Interview' }
    post:
      tags: [interviews]
      summary: Schedule or reschedule an interview (upsert per candidate + round)
      description: |
        Transitions candidate stage: Sourced → R1 Scheduled (round=1) or
        R1 Complete → R2 Scheduled (round=2). Re-scheduling clears any prior
        cancellation on the same row.
      responses:
        '201': { description: Created or upserted }
        '400': { description: Invalid transition or validation failure }
  /api/interviews/{id}:
    get:
      tags: [interviews]
      summary: Get a single interview
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: OK }
    patch:
      tags: [interviews]
      summary: Record Select / Reject / No-show — transitions candidate to R1/R2 Complete
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [result]
              properties:
                result: { type: string, enum: [Select, Reject, No-show] }
      responses:
        '200': { description: OK }
        '400': { description: Invalid transition or validation failure }
    delete:
      tags: [interviews]
      summary: Cancel a scheduled interview (soft delete + revert candidate stage)
      description: |
        Atomic soft-cancel: sets cancelled_at on the interview row AND walks the
        candidate's pipeline stage back one step (R1 Scheduled → Sourced;
        R2 Scheduled → R1 Complete) in a single transaction. Cancelling a
        completed interview (one with a result) returns 400 — the audit trail
        of what actually happened wins over a soft delete.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: reason, in: query, schema: { type: string, maxLength: 500 }, description: Free-text reason ("interviewer sick", "candidate withdrew", etc.) }
      responses:
        '200': { description: Updated interview row with cancelledAt populated }
        '400': { description: Already-completed interview, or invalid id }
        '404': { description: No interview with that id }
...